+++ doc/releaseNotes.md
@@ -183,4 +183,5 @@ This file describes the main feature changes for each readout.exe released versi
 ## next
 - Updated configuration parameters:
   - added parameters for receiverFMQ: dumpRDH, dumpTF
+  - added readout.aggregatorSliceTimeout
 - readRaw.exe: fix reading large files, realign was not done for last page.